[refactor]:Socket #42 package 구조 refactoring
