+++ SocketService/src/main/java/com/kusitms/socketservice/global/config/SocketSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.kusitms.socketservice.config;
+package com.kusitms.socketservice.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.messaging.MessageSecurityMetadataSourceRegistry;
