+++ SocketService/src/main/java/com/kusitms/socketservice/global/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.kusitms.socketservice.config;
+package com.kusitms.socketservice.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
