+++ SocketService/src/main/java/com/kusitms/socketservice/global/config/auth/UserAuthentication.java
@@ -1,4 +1,4 @@
-package com.kusitms.socketservice.config.auth;
+package com.kusitms.socketservice.global.config.auth;
 
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.GrantedAuthority;
