+++ SocketService/src/main/java/com/kusitms/socketservice/global/error/httpException/UnauthorizedException.java
@@ -1,9 +1,11 @@
-package com.kusitms.socketservice.error.socketException;
+package com.kusitms.socketservice.global.error.httpException;
 
-import com.kusitms.socketservice.error.ErrorCode;
+import com.kusitms.socketservice.global.error.ErrorCode;
 
 public class UnauthorizedException extends BusinessException {
+
     public UnauthorizedException(ErrorCode errorCode) {
         super(errorCode);
     }
 }
+
