+++ SocketService/src/main/java/com/kusitms/socketservice/global/error/socketException/BusinessException.java
@@ -1,6 +1,6 @@
-package com.kusitms.socketservice.error.socketException;
+package com.kusitms.socketservice.global.error.socketException;
 
-import com.kusitms.socketservice.error.ErrorCode;
+import com.kusitms.socketservice.global.error.ErrorCode;
 import lombok.Getter;
 import org.springframework.messaging.MessageDeliveryException;
