+++ SocketService/src/main/java/com/kusitms/socketservice/global/error/socketException/UnauthorizedException.java
@@ -1,11 +1,9 @@
-package com.kusitms.socketservice.error.httpException;
+package com.kusitms.socketservice.global.error.socketException;
 
-import com.kusitms.socketservice.error.ErrorCode;
+import com.kusitms.socketservice.global.error.ErrorCode;
 
 public class UnauthorizedException extends BusinessException {
-
     public UnauthorizedException(ErrorCode errorCode) {
         super(errorCode);
     }
 }
-
